chenged type of radius to double

+++ FlowerBed.java
@@ -10,7 +10,7 @@ public class FlowerBed {
 				new InputStreamReader(System.in));
 		
 		System.out.println("Input radius (in meters) of flower bed:");
-		int r = Integer.parseInt(br.readLine());
+		double r = Double.parseDouble(br.readLine());
 		
 		System.out.println("Perimeter is " + (2*3.14*r) + " m");
 		System.out.println("Area is " + (3.14*r*r) + " m^2");
